refactored `FileBrowserModel.normalizePath` => `Contents.IManager.resolvePath`
